requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pillow
pytesseract
pymupdf
python-dotenv
aiohttp
tenacity
PyPDF2
urllib3
requests
pypdf